+++ deno/payloads/v9/template.ts
@@ -3,6 +3,7 @@
  */
 
 import type { Snowflake } from '../../globals.ts';
+import type { Locale } from '../../rest/common.ts';
 import type { RESTPostAPIGuildsJSONBody } from '../../rest/v9/mod.ts';
 import type { APIUser } from './user.ts';
 
@@ -60,6 +61,6 @@ export interface APITemplate {
 
 export interface APITemplateSerializedSourceGuild extends Omit<RESTPostAPIGuildsJSONBody, 'icon'> {
 	description: string | null;
-	preferred_locale: string;
+	preferred_locale: Locale;
 	icon_hash: string | null;
 }